added PhySG renderer & README: images/demo_render_PhySG_2.png

+++ lib/utils_io.py
@@ -244,7 +244,7 @@ def read_cam_params(camFile: Path) -> list:
     #     camNum = int(camIn.readline().strip() )
         cam_data = camIn.read().splitlines()
     cam_num = int(cam_data[0])
-    cam_params = np.array([x.split(' ') for x in cam_data[1:]]).astype(np.float)
+    cam_params = np.array([x.split(' ') for x in cam_data[1:]]).astype(np.float32)
     assert cam_params.shape[0] == cam_num * 3
     cam_params = np.split(cam_params, cam_num, axis=0) # [[origin, lookat, up], ...]
     return cam_params
